% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amre_abundance}
\alias{amre_abundance}
\title{American Redstart population relative abundance}
\format{
A tibble with 10 rows, 2 columns:
\describe{
\item{Population}{ID of the node}
\item{Relative_abundance}{The relative abundance of the node}
}
}
\usage{
amre_abundance
}
\description{
The tibble provides relative abundance for all nodes (breeding and nonbreeding)
}
\keyword{datasets}
