% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amre_conn}
\alias{amre_conn}
\title{American Redstart connectivity}
\format{
A tibble with 3 rows, 3 columns:
\describe{
\item{1}{Connectivity of nonbreeding population ALM with breeding populations WB, NT, and ST}
\item{2}{Connectivity of nonbreeding population CAR with breeding populations WB, NT, and ST}
\item{3}{Connectivity of nonbreeding population AONU with breeding populations WB, NT, and ST}
}
}
\usage{
amre_conn
}
\description{
The matrix provides the connectivity output from the JAGS model with \code{amre_assign} and \code{amre_abundance} data
}
\keyword{datasets}
