% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amre_nonbreeding_consistency}
\alias{amre_nonbreeding_consistency}
\title{Assignment probabilities for the American Redstart nonbreeding data}
\format{
A tibble with 135 rows, 7 columns:
\describe{
\item{Sample}{Individual identifier}
\item{WB}{Western Boreal assignment probability}
\item{BR}{Basin Rockies assignment probability}
\item{NT}{Northern Temperate assignment probability}
\item{ST}{Southern Temperate assignment probability}
\item{MP}{Maritime Provinces assignment probability}
\item{Nonbreeding_region}{Conservation region of the sampling site}
}
}
\usage{
amre_nonbreeding_consistency
}
\description{
The tibble provides assignment probabilities the individuals sampled from the nonbreeding range.
Assignment probabilities are provided from the DeSaix et al., 2023 dataset using the assignment consistency.
}
\keyword{datasets}
