% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amre_nonbreeding_data}
\alias{amre_nonbreeding_data}
\title{Metadata for the American Redstart samples from the stationary nonbreeding range}
\format{
A tibble with 135 rows, 6 columns:
\describe{
\item{Sample}{Individual identifier}
\item{Site}{Nonbreeding sampling site}
\item{Nonbreeding_region}{Conservation region of the sampling site}
\item{Lat}{Latitude of nonbreeding site}
\item{Lon}{Longitude of nonbreeding site}
\item{Breeding_assignment}{Breeding population the sample was assigned to}
}
}
\usage{
amre_nonbreeding_data
}
\description{
The tibble provides metadata for all of the individuals sampled from the nonbreeding range.
Each individual has a unique identifier, location data, nonbreeding sampling coordinates, and breeding population assignment from the WGSassign software
}
\keyword{datasets}
