% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_genoscape_overlap.R
\name{check_genoscape_overlap}
\alias{check_genoscape_overlap}
\title{Check which genoscape polygons overlap}
\usage{
check_genoscape_overlap(x)
}
\arguments{
\item{x}{A polygon sf object for all genoscape clusters}
}
\value{
List of genoscape clusters that overlap
}
\description{
Check which genoscape polygons overlap
}
