% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comyel_abd_season_proj}
\alias{comyel_abd_season_proj}
\title{Seasonal abundance of the Common Yellowthroat}
\format{
A RasterStack with abundance for 4 stages: nonbreeding, prebreeding_migration, breeding, and postbreeding_migration
}
\source{
\url{https://ebird.org/science/status-and-trends}
}
\usage{
comyel_abd_season_proj
}
\description{
A raster with different seasonal layers of abundance for the Common Yellowthroat
base on ebirdst data.
}
\keyword{datasets}
