% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comyel_breed_smooth}
\alias{comyel_breed_smooth}
\title{Breeding range of the Common Yellowthroat}
\format{
A multipolygon with 1 feature and 2 fields:
\describe{
\item{season}{breeding}
\item{layer}{range}
}
}
\source{
\url{https://ebird.org/science/status-and-trends/comyel}
}
\usage{
comyel_breed_smooth
}
\description{
A polygon of the breeding range based on the seasonal abundance data
}
\keyword{datasets}
