% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comyel_wintering_data}
\alias{comyel_wintering_data}
\title{Metadata for the Common Yellowthroat wintering samples}
\format{
A tibble with 106 rows, 5 columns:
\describe{
\item{Sample}{Individual identifier}
\item{Location}{Wintering site}
\item{Lat}{Latitude of wintering site}
\item{Long}{Longitude of wintering site}
\item{BreedingAssignment}{Breeding population the sample was assigned to with Rubias}
}
}
\usage{
comyel_wintering_data
}
\description{
The tibble provides metadata for all of the individuals sampled from the wintering range.
Each individual has a unique identifier, location data, winter sampling coordinates, and breeding population assignment from Rubias.
}
\keyword{datasets}
