% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions_simple}
\alias{ecoregions_simple}
\title{Ecoregions of the western hemisphere}
\format{
A simple feature collection with 27 features (ecoregions) and 8 fields
\describe{
}
}
\usage{
ecoregions_simple
}
\description{
These polygons were simplified with rmapshaper::ms_simplify()
}
\keyword{datasets}
