% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abundance_stats.R
\name{get_abundance_stats}
\alias{get_abundance_stats}
\title{Get abundance summaries for the different nodes}
\usage{
get_abundance_stats(range_raster, node_poly, node_names)
}
\arguments{
\item{range_raster}{SpatRaster of seasonal abundance for the range of interest}

\item{node_poly}{Polygons of the nodes}

\item{node_names}{Names of the nodes}
}
\value{

}
\description{
Get abundance summaries for the different nodes
}
