% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FT_fit.R
\name{get_FT_fit}
\alias{get_FT_fit}
\title{Evaluate goodness of fit (Freeman-Tukey statistic)}
\usage{
get_FT_fit(network_model)
}
\arguments{
\item{network_model}{Network model output from mignette::run_network_model()}
}
\value{
Returns a ggplot of the observed vs. simulated Freeman-Tukey distributions, and Bayesian pvalue
}
\description{
Evaluate goodness of fit (Freeman-Tukey statistic)
}
