% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genoscape_abunds.R
\name{get_genoscape_abunds}
\alias{get_genoscape_abunds}
\title{Get relative abundance for genoscape}
\usage{
get_genoscape_abunds(genoscape, abunds, genoscape_names = names(genoscape))
}
\arguments{
\item{genoscape}{SpatRaster object of genoscape layers with q-values}

\item{abunds}{SpatRaster object of ebirdst abundance}

\item{genoscape_names}{Character vector of genoscape layer names manually provided. By default the function obtains this from \code{names(genoscape)}.}
}
\value{
A two-column matrix with population names in the first column and relative abundance values in the second column
}
\description{
Get relative abundance for genoscape
}
