% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jags_data.R
\name{get_jags_data}
\alias{get_jags_data}
\title{Set up the data for the migratory network model in JAGS}
\usage{
get_jags_data(abundance, assignment, bnode_names, wnode_names)
}
\arguments{
\item{abundance}{Tibble of abundance data across populations (model nodes)}

\item{assignment}{Tibble of assignment among the populations}

\item{bnode_names}{Names of breeding populations/nodes}

\item{wnode_names}{Names of nonbreeding populations/nodes}
}
\value{
A list of of input data for the JAGS model
}
\description{
Set up the data for the migratory network model in JAGS
}
