% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jags_model.R
\name{get_jags_model}
\alias{get_jags_model}
\title{Create the migratory network model to be run in JAGS}
\usage{
get_jags_model(base_filename = "jags_genetic_model", model_type = 1)
}
\arguments{
\item{base_filename}{Character string of file name for the model .txt file that will be saved.}

\item{model_type}{Integer value, where 1 indicates genetic data being used, 2 indicates genetic and geolocator data}
}
\value{
Returns the file name of the model and saves the model as a text file
}
\description{
Create the migratory network model to be run in JAGS
}
