% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range_smooth.R
\name{get_range_smooth}
\alias{get_range_smooth}
\title{Convert seasonable abundance raster to shapefile}
\usage{
get_range_smooth(
  abd_season_proj,
  species,
  ne_land,
  pred_region,
  split_migration = FALSE,
  show_yearround = FALSE
)
}
\arguments{
\item{abd_season_proj}{Name of seasonal abundance raster (projected)}

\item{species}{Six-letter code for the bird species from ebirdst}

\item{ne_land}{Natural Earth land data (same projection as abd_season_proj)}

\item{pred_region}{Predicted region}

\item{split_migration}{TRUE/FALSE of whether to split spring and fall migration}

\item{show_yearround}{TRUE/FALSE of whether to create shapefile for year round abundance}
}
\value{
A shapefile
}
\description{
Convert seasonable abundance raster to shapefile
}
