% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_abunds.R
\name{get_raster_abunds}
\alias{get_raster_abunds}
\title{Get relative abundance for genoscape}
\usage{
get_raster_abunds(populations, abunds, pop_names = names(populations))
}
\arguments{
\item{populations}{SpatRaster object of layers with values delineating membership probabilities to a population}

\item{abunds}{SpatRaster object of ebirdst abundance}

\item{pop_names}{Character vector of raster layer names manually provided. By default the function obtains this from \code{names(populations)}.}
}
\value{
A two-column matrix with population names in the first column and relative abundance values in the second column
}
\description{
Get relative abundance for genoscape
}
