% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vector_abunds.R
\name{get_vector_abunds}
\alias{get_vector_abunds}
\title{Get relative abundance for vector files}
\usage{
get_vector_abunds(populations, abunds, pop_names = populations[[1]][[1]])
}
\arguments{
\item{populations}{SpatVector object of population boundaries}

\item{abunds}{SpatRaster object of ebirdst abundance}

\item{pop_names}{Character vector of population names manually provided. By default the function obtains this from \code{populations[[1]][[1]]}.}
}
\value{
A two-column matrix with population names in the first column and relative abundance values in the second column
}
\description{
Get relative abundance for vector files
}
