% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genoscape_overlap.R
\name{plot_genoscape_overlap}
\alias{plot_genoscape_overlap}
\title{Plot genoscape overlap}
\usage{
plot_genoscape_overlap(x, fill)
}
\arguments{
\item{x}{A polygon sf object for all genoscape clusters}

\item{fill}{A character vector of colors}
}
\value{
A ggplot with the overlap regions in black
}
\description{
Plot genoscape overlap
}
