% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_smooth.R
\name{range_smooth}
\alias{range_smooth}
\title{Convert seasonable abundance raster to shapefile}
\usage{
range_smooth(
  abd_season,
  ne_land,
  smooth_area = 1000,
  split_migration = FALSE,
  show_yearround = FALSE
)
}
\arguments{
\item{abd_season}{Name of seasonal abundance RasterStack (projected)}

\item{ne_land}{Natural Earth land data (same projection as abd_season)}

\item{smooth_area}{km^2 region to smooth}

\item{split_migration}{TRUE/FALSE of whether to split spring and fall migration}

\item{show_yearround}{TRUE/FALSE of whether to create shapefile for year round abundance}
}
\value{
Sf object for all seasons
}
\description{
Convert seasonable abundance raster to shapefile
}
