% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_network_model.R
\name{run_network_model}
\alias{run_network_model}
\title{Set up the data for the migratory network model for JAGS}
\usage{
run_network_model(
  abundance,
  brnode_names,
  nbnode_names,
  model,
  base_filename = "jags",
  nb2br_assign = NULL,
  br2nb_assign = NULL,
  iter.increment = 5e+05,
  n.thin = 4,
  n.burnin = 1e+05,
  n.chains = 2,
  parallel = FALSE
)
}
\arguments{
\item{abundance}{Tibble of abundance data across populations (model nodes)}

\item{brnode_names}{Character vector of names of breeding populations/nodes}

\item{nbnode_names}{Character vector of names of nonbreeding populations/nodes}

\item{model}{Character string of assignment model type ("BR" = encounter nonbreeding/recovery breeding, NB = encounter breeding/recovery nonbreeding, FULL = both types of data provided)}

\item{base_filename}{Character string of file name for the model .txt file that will be saved. Model type added as suffix to name.}

\item{nb2br_assign}{Tibble of assignment for encounter season as nonbreeding (columns) and recovery season as breeding populations (rows)}

\item{br2nb_assign}{Tibble of assignment for encounter season as breeding (columns) to recovery season as nonbreeding populations (rows)}

\item{iter.increment}{Integer value of jagsUI::autojags() parameter iter.increment}

\item{n.thin}{Integer value of jagsUI::autojags() parameter n.thin}

\item{n.burnin}{Integer value of jagsUI::autojags() parameter n.burnin}

\item{n.chains}{Integer value of jagsUI::autojags() parameter n.chains}

\item{parallel}{Logical value of jagsUI::autojags() paramter parallel}
}
\value{
A list of 1) "conn" is a dataframe of the connectivity matrix with node names, 2) "jags_out" is a list of the full standard jagsUI::autojags() output
}
\description{
Set up the data for the migratory network model for JAGS
}
