% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scape_to_shape.R
\name{scape_to_shape}
\alias{scape_to_shape}
\title{Convert genoscape raster stack to polygons}
\usage{
scape_to_shape(x, prob_threshold = 0.5, d = 1000, f = 1000, s = 3)
}
\arguments{
\item{x}{Genoscape RasterStack}

\item{prob_threshold}{Probability value to include raster cell in polygon}

\item{d}{Distance threshold for smoothr::drop_crumbs()}

\item{f}{Distance threshold for smoothr::fill_holes()}

\item{s}{Smoothness}
}
\value{
A polygon sf object for all genoscape clusters
}
\description{
Convert genoscape raster stack to polygons
}
