% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_points.R
\name{snap_points}
\alias{snap_points}
\title{Snap points to polygons}
\usage{
snap_points(x, y, max_dist = 1000)
}
\arguments{
\item{x}{Points to snap}

\item{y}{Polygons to snap to}

\item{max_dist}{Distance threshold}
}
\value{
Returns points snapped to polygons
}
\description{
Snap points to polygons
}
