% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_conn.R
\name{tidy_conn}
\alias{tidy_conn}
\title{Prepare a tibble for plotting credible intervals}
\usage{
tidy_conn(conn_mat, statistic, brnode_names, nbnode_names)
}
\arguments{
\item{conn_mat}{Matrix of data}

\item{statistic}{Character string specifying statistic of interest}

\item{brnode_names}{Character vector of breeding node names}

\item{nbnode_names}{Character vector of nonbreeding node names}
}
\value{
Returns a long tibble
}
\description{
Prepare a tibble for plotting credible intervals
}
