% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amre_conn}
\alias{amre_conn}
\title{American Redstart connectivity}
\format{
A matrix with 5 rows, 5 columns:
\describe{
\item{ALM}{Connectivity of nonbreeding population ALM with the breeding populations}
\item{LCA}{Connectivity of nonbreeding population LCA with the breeding populations}
\item{HCA}{Connectivity of nonbreeding population HCA with the breeding populations}
\item{CAR}{Connectivity of nonbreeding population CAR with the breeding populations}
\item{AONU}{Connectivity of nonbreeding population AONU with the breeding populations}
}
}
\usage{
amre_conn
}
\description{
The matrix provides the connectivity output from the JAGS model with \code{amre_assign} and \code{amre_abundance} data
}
\keyword{datasets}
