% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ebirdst_abd_season.R
\name{get_ebirdst_abd_season}
\alias{get_ebirdst_abd_season}
\title{Get seasonal abundance data from ebirdst}
\usage{
get_ebirdst_abd_season(species, force = FALSE)
}
\arguments{
\item{species}{Six-letter code for the bird species from ebirdst}

\item{force}{TRUE/FALSE of redownloading species data from ebirdst}
}
\value{
A raster stack of abundance data by season
}
\description{
Get seasonal abundance data from ebirdst
}
