% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jags_data.R
\name{get_jags_data}
\alias{get_jags_data}
\title{Set up the data for the migratory network model in JAGS}
\usage{
get_jags_data(
  abundance,
  bnode_names,
  wnode_names,
  model,
  nb2br_assign = NULL,
  br2nb_assign = NULL
)
}
\arguments{
\item{abundance}{Tibble of abundance data across populations (model nodes)}

\item{bnode_names}{Names of breeding populations/nodes}

\item{wnode_names}{Names of nonbreeding populations/nodes}

\item{model}{Select integer value of assignment model type (1 = known nonbreeding/inferred breeding, 2 = known breeding/inferred nonbreeding, 3 = both types of data provided)}

\item{nb2br_assign}{Tibble of assignment for known nonbreeding (columns) to inferred breeding populations (rows)}

\item{br2nb_assign}{Tibble of assignment for known breeding (columns) to inferred nonbreeding populations (columns)}
}
\value{
A list of of input data for the JAGS model
}
\description{
Set up the data for the migratory network model in JAGS
}
