% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jags_model.R
\name{get_jags_model}
\alias{get_jags_model}
\title{Create the migratory network model to be run in JAGS}
\usage{
get_jags_model(base_filename = "jags", model = 1)
}
\arguments{
\item{base_filename}{Character string of file name for the model .txt file that will be saved. Model integer added as suffix to name.}

\item{model}{Select integer value of assignment model type (1 = known nonbreeding/inferred breeding, 2 = known breeding/inferred nonbreeding, 3 = both types of data provided)}
}
\value{
Returns the file name of the model and saves the model as a text file
}
\description{
Create the migratory network model to be run in JAGS
}
