% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_create.R
\name{net_create}
\alias{net_create}
\title{Create the network for visualizing from the JAGS model}
\usage{
net_create(
  c_matrix,
  node.types = c("B", "W"),
  node.names = list(),
  margin = 0.05,
  connected_tol = 0.001
)
}
\arguments{
\item{c_matrix}{Matrix of network connectivity. This is the output from the JAGS run}

\item{node.types}{Node types of seasonal stages of annual cycle: "B" for breeding, "W" for wintering/nonbreeding}

\item{node.names}{Node names of the connectivity model. These need to be in the same order of the JAGS run.}

\item{margin}{Float value}

\item{connected_tol}{Float value}
}
\value{
Returns a network
}
\description{
Create the network for visualizing from the JAGS model
}
