% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_draw.R
\name{net_draw}
\alias{net_draw}
\title{Visualize the network from the JAGS model}
\usage{
net_draw(net_object)
}
\arguments{
\item{net_object}{Net object produced from \code{net_create}}
}
\value{
Returns a ggplot of the network
}
\description{
Visualize the network from the JAGS model
}
