% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_network_model.R
\name{run_network_model}
\alias{run_network_model}
\title{Set up the data for the migratory network model for JAGS}
\usage{
run_network_model(
  abundance,
  bnode_names,
  wnode_names,
  model,
  base_filename = "jags",
  nb2br_assign = NULL,
  br2nb_assign = NULL,
  iter.increment = 5e+05,
  n.thin = 4,
  n.burnin = 1e+05,
  n.chains = 2,
  parallel = FALSE
)
}
\arguments{
\item{abundance}{Tibble of abundance data across populations (model nodes)}

\item{bnode_names}{Names of breeding populations/nodes}

\item{wnode_names}{Names of nonbreeding populations/nodes}

\item{model}{Select integer value of assignment model type (1 = encounter nonbreeding/recovery breeding, 2 = ecounter breeding/recovery nonbreeding, 3 = both types of data provided)}

\item{base_filename}{Character string of file name for the model .txt file that will be saved. Model integer added as suffix to name.}

\item{nb2br_assign}{Tibble of assignment for encounter season as nonbreeding (columns) and recovery season as breeding populations (rows)}

\item{br2nb_assign}{Tibble of assignment for encounter season as breeding (columns) to recovery season as nonbreeding populations (rows)}

\item{iter.increment}{Integer value of jagsUI::autojags() parameter iter.increment}

\item{n.thin}{Integer value of jagsUI::autojags() parameter n.thin}

\item{n.burnin}{Integer value of jagsUI::autojags() parameter n.burnin}

\item{n.chains}{Integer value of jagsUI::autojags() parameter n.chains}

\item{parallel}{Logical value of jagsUI::autojags() paramter parallel}
}
\value{
A list of 1) "conn_g" is a dataframe of the connectivity matrix with node names, 2) "jags_out" is a list of the full standard jagsUI::autojags() output
}
\description{
Set up the data for the migratory network model for JAGS
}
